# resources/openapi.yaml
openapi: 3.0.3
info:
  title: PoisaPay REST API
  version: "1.0.0"
  description: >
    PoisaPay stablecoin payment platform API. Bearer-token (Sanctum) authenticated;
    money-moving endpoints accept an Idempotency-Key header and are rate-limited.
servers:
  - url: /api/v1
    description: Version 1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Collapses retries of a money-moving request into one operation.
  schemas:
...
    Address:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        address: { type: string }
        status: { type: string, enum: [active, pending, blocked] }
        whitelisted: { type: boolean }
        cooldown_until: { type: string, format: date-time, nullable: true }
security:
  - bearerAuth: []
paths:
...
  /security/addresses:
    get:
      tags: [Security]
      summary: List withdrawal-address whitelist entries
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Address' } } } } } } }
    post:
      tags: [Security]
      summary: Add a withdrawal address (enters cooldown)
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address: { type: string }
                label: { type: string }
                chain_id: { type: integer }
      responses:
        "201": { description: Created }
  /security/addresses/{id}:
    delete:
      tags: [Security]
      summary: Remove a saved withdrawal address
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        "200": { description: OK }
...